// client/src/app/Models/plant-type.ts
export interface PlantType {
  id?: number;
  Id?: number;
  commonName?: string;
  CommonName?: string;
  scientificName?: string;
  ScientificName?: string;
  wateringFrequencyDays?: number;
  WateringFrequencyDays?: number;
  fertilizingFrequencyDays?: number;
  FertilizingFrequencyDays?: number;
  sunlightNeeds?: string;
  SunlightNeeds?: string;
}